requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
Deprecated==1.2.13
dj-database-url==0.5.0
Django==2.2.10
django-cors-headers==3.8.0
django-phone-verify==2.0.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
mysqlclient==2.0.3
nexmo==2.5.2
phonenumbers==8.12.33
Pillow==8.3.2
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.0.0
Unipath==1.1
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.12.1

